Tries to fix things but probably break them
I don't actually know go so it is very possible that I completely messed up the 1/2 line addition. Please make sure I did not.
Long story short the check for dirContainsRegex function when it tried to access a file that was already open in another process. Because the function would terminate upon encountering an error, this means certain files would not be checked. This is very sad. I have made it less sad.