Fix dirContainsRegex (#86)
* Very useful commit

Fixed one whole typo.

* Tries to fix things but probably break them

I don't actually know go so it is very possible that I completely messed up the 1/2 line addition. Please make sure I did not.
Long story short the check for dirContainsRegex function when it tried to access a file that was already open in another process. Because the function would terminate upon encountering an error, this means certain files would not be checked. This is very sad. I have made it less sad.

* I'm an idiot

It's strings, not string.

* Nicer error handling

I'm not sure if I did the error handling correctly, but I sure hope I did.

Co-authored-by: Bryant Jin <18301794+Drakonhuntas@users.noreply.github.com>